optimize preload.js
