+++ src/main.js
@@ -39,7 +39,7 @@
         }
     });
 
-    //win?.removeMenu();
+    win?.removeMenu();
     win?.loadFile(path.join(__dirname, './pages/home.html'));
 
     // The default window background color
