+++ src/pages/settings.html
@@ -236,7 +236,7 @@
                         </h3>
 
                         <p class="settings_description">
-                            بداء تشغيل التطبيق مع بداية تشغيل النظام, يدعم نظام ويندوز فقط
+                            بدأ التطبيق مع بداية النظام, يدعم نظام ويندوز فقط
                         </p>
                     </div>
 
@@ -259,7 +259,7 @@
                         </h3>
 
                         <p class="settings_description">
-                            بداء تشغيل التطبيق بدون واجه رسومية gui
+                            بدأ التطبيق بدون واجهة رسومية بالخلفية
                         </p>
                     </div>
 
@@ -364,7 +364,6 @@
             </ul>
 
             <button id="save">حفظ</button>
-
         </div>
 
     </div>
